// src/models/Weather.ts
export type WeatherConditionType =
  | 'clear' //ясно
  | 'partly-cloudy' //малооблачно
  | 'cloudy' //облачно с прояснениями
  | 'overcast' //пасмурно
  | 'drizzle' //морось
  | 'light-rain' //небольшой дождь
  | 'rain' //дождь
  | 'moderate-rain' //умеренно сильный дождь
  | 'heavy-rain' //сильный дождьs
  | 'continuous-heavy-rain' //длительный сильный дождь
  | 'showers' //ливень
  | 'wet-snow' //дождь со снегом
  | 'light-snow' // небольшой снег
  | 'snow' //снег
  | 'snow-showers' //снегопад
  | 'hail' //град
  | 'thunderstorm' //гроза
  | 'thunderstorm-with-rain' //дождь с грозой
  | 'thunderstorm-with-hail'; //гроза с градом